service/backend/service/fastapi-server/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.7
psycopg2-binary>=2.9.5
pydantic>=1.10.7
passlib>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
bcrypt>=4.0.1
python-dotenv>=1.0.0
httpx>=0.24.0
openai>=1.0.0
pandas>=2.0.0
langchain>=0.1.0
langchain_core>=0.1.0
langchain_openai>=0.0.5
langchain-community>=0.0.1
langsmith>=0.0.60
python-jose[cryptography]>=3.3.0 